Moving the logic for the connection button to a new component
